+++ src/StateSmith.Cli/Run/Versionable.cs
@@ -1,0 +1,6 @@
+﻿namespace StateSmith.Cli.Run;
+
+public class Versionable
+{
+    public SemVersion version = new(0, 1, 0);
+}