+++ src/StateSmith.Cli/SemVersion.cs
@@ -1,0 +1,16 @@
+﻿namespace StateSmith.Cli;
+
+public class SemVersion
+{
+    public int Major { get; set; }
+    public int Minor { get; set; }
+    public int Patch { get; set; }
+    public string tag { get; set; } = "";
+
+    public SemVersion(int major, int minor, int patch)
+    {
+        Major = major;
+        Minor = minor;
+        Patch = patch;
+    }
+}